requirements.txt
python-docx
Pillow==9.3.0